requirements.txt
Flask==2.3.3
flask-cors==6.0.1
python-dotenv==0.21.0
functions-framework==3.1.0
pyjwt==2.4.0
bcrypt==4.3.0
requests==2.32.5